Fixes for adding support for ElasticSearch v8 (#216)
* Fix issue with index not created from tutorial with ElasticSearch v8

With ElasticSearch v8, whenever a new index needs to be created, the
request should not contain an invalid `Accept` header.

As of now, the `Accept` header is passed from the browser and
accordingly passed on to the elastic API call which in order raises an
issue.

The value for `Accept` passed from the browser is `*/*` but Elastic
doesn't like that. Added a fix for that issue by removing the passed
`Accept` header being forwarded.

* Allow default index creation for version 8 as well

The default index created is used by the dashboard to make subsequent
requests. As of now, the code supported Elastic v7 and so while using
with Elastic v8, it was failing since the default index was not being
created.

Added a fix for that.