Use ES Client to get Index Mapping call (#226)
* Add changes to the index mapping get method to use the client for calls

The es7 client exposes a method that can be used to make direct calls
to a certain endpoint. Advantage of using that is that we can have
automatic retries in case the call fails for the first time. This is
useful in cases where Elastic takes a while to startup or so.

Added changes to start using that method in order to make calls to get
the index mapping.

* Add a fix to format the index name in the URL

* Remove ESUrl from the index name endpoint since it's added by client