Add support to create template with index template
Since ES v7.9, index templates were introduced that made the older
templates legacy and started throwing errors.

This commit changes that step to create the template as an index
template. If it fails, as a fallback the template is tried to be created
as a legacy template since the ES version might be older than 7.9 and
there is no way to determine that.